+++ index.html
@@ -226,6 +226,8 @@
       </div>
     </section>
   </template>
-<script src="js/functions.js"></script>
+<!-- <script type="module" src="/js/functions.js"></script> -->
+<script src="/js/functions.js"></script>
+<script type="module" src="/js/main.js"></script>
 </body>
 </html>
