+++ js/main.js
@@ -1,0 +1,2 @@
+import {CreatePosts} from '/js/data.js';
+
